+++ examples/my_app/lib/my_app/users/index_response.ex
@@ -1,5 +1,5 @@
 defmodule MyApp.Users.JSONIndexResponse do
-  use Valspec
+  use Valspec.Schema
 
   valspec_schema do
     embeds_many(:data, MyApp.Users.UserSchema)
