+++ test/valspec_test.exs
@@ -1,8 +1,4 @@
 defmodule ValspecTest do
   use ExUnit.Case
   doctest Valspec
-
-  test "greets the world" do
-    assert Valspec.hello() == :world
-  end
 end
